+++ src/pages/Dashboard/ArNSStatsPanel.tsx
@@ -1,13 +1,37 @@
 import Placeholder from '@src/components/Placeholder';
+import Tooltip from '@src/components/Tooltip';
 import useArNSStats from '@src/hooks/useArNSStats';
 import { formatWithCommas } from '@src/utils';
+import { InfoIcon } from 'lucide-react';
 
 const ArNSStatsPanel = () => {
   const { data: arnsStats } = useArNSStats();
 
   return (
     <div className="flex min-w-[22rem] flex-col rounded-xl border border-grey-500 px-6 py-5">
-      <div className=" text-sm text-mid">ArNS Names Purchased</div>
+      <div className="flex items-center">
+        <div className="grow text-sm text-mid">ArNS Names Purchased</div>
+
+        {arnsStats && (
+          <Tooltip
+            message={
+              <div>
+                <div className="mb-2 font-semibold text-high">ArNS Stats</div>
+                <div>
+                  <div>Active Names: {arnsStats.totalActiveNames}</div>
+                  <div>Returned Names: {arnsStats.totalReturnedNames}</div>
+                  <div>
+                    Grace Period Names: {arnsStats.totalGracePeriodNames}
+                  </div>
+                  <div>Reserved Names: {arnsStats.totalReservedNames}</div>
+                </div>
+              </div>
+            }
+          >
+            <InfoIcon className="size-[1.125rem] text-low" />
+          </Tooltip>
+        )}
+      </div>
       <div className="self-center px-24 py-6 text-center text-[2.625rem]">
         {arnsStats ? (
           formatWithCommas(arnsStats.namesPurchased)